pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "solana-swarm"
version = "0.1.0"
description = "Solana Swarm Intelligence Framework for AI-powered trading strategies"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Solana Swarm Team", email = "team@solana-swarm.com"}
]
keywords = ["solana", "ai", "trading", "defi", "agents", "swarm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business :: Financial :: Investment",
]
requires-python = ">=3.8"
dependencies = [
    "pydantic>=2.0.0",
    "aiohttp>=3.8.0",
    "solana>=0.30.0",
    "solders>=0.18.0",
    "spl-token>=0.11.0",
    "PyYAML>=6.0",
    "click>=8.0.0",
    "asyncio-mqtt>=0.11.0",
    "websockets>=11.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
docs = [
    "mkdocs>=1.4.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.20.0",
]

[project.urls]
Homepage = "https://github.com/solana-swarm/solana-swarm"
Documentation = "https://docs.solana-swarm.com"
Repository = "https://github.com/solana-swarm/solana-swarm"
Issues = "https://github.com/solana-swarm/solana-swarm/issues"

[project.scripts]
solana-swarm = "solana_swarm.cli.main:app"
swarm-chat = "solana_swarm.cli.chat:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["solana_swarm*"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=solana_swarm --cov-report=html --cov-report=term"
asyncio_mode = "auto"
